added IClassFixture and IDisposable
