+++ BALTest/Services/UserServiceTests.cs
@@ -9,54 +9,23 @@ using System.Collections.Generic;
 
 namespace BALTest.Services
 {
-    public class UserServiceTests
+    public class UserServiceTests : IClassFixture<DataFixture>
     {
-        private static DbContextOptions<DbSportsBuzzContext> dbContextOptions = new DbContextOptionsBuilder<DbSportsBuzzContext>()
-                  .UseInMemoryDatabase(databaseName: "db_SportsBuzz")
-                  .Options;
-        DbSportsBuzzContext context;
+        DataFixture _fixture;
         UserService userService;
         Mock<IEncrypt> encrypt;
 
-        public UserServiceTests()
+        public UserServiceTests(DataFixture fixture)
         {
-            context = new DbSportsBuzzContext(dbContextOptions);
-            context.Database.EnsureCreated();
+            _fixture = fixture;
             encrypt = new Mock<IEncrypt>();
-            userService = new UserService(context, encrypt.Object);
-        }
-
-        public void SeedDatabase()
-        {
-            var role = new List<TblUserRole>()
-            {
-                new TblUserRole(){ UserRoleId=1, UserRole="Team Manager"  },
-                new TblUserRole(){ UserRoleId=2, UserRole="Ground Manager" }
-            };
-            context.TblUserRoles.AddRange(role);
-            context.SaveChanges();
-
-            var user = new List<TblUser>()
-            {
-                new TblUser(){ UserId=1, FirstName="Gokul", LastName="B V", Email="bvgokulgok@gmail.com", PhoneNum=8834834383, Password="1234", UserRole=1, CreatedDate=DateTime.Now, UpdatedDate=null, Active=true },
-                new TblUser(){ UserId=2, FirstName="Gokul", LastName="B V", Email="bvgokul@gmail.com", PhoneNum=8834834383, Password="1234", UserRole=2, CreatedDate=DateTime.Now, UpdatedDate=null, Active=true },
-                new TblUser(){ UserId=3, FirstName="Gokul", LastName="B V", Email="bvgok@gmail.com", PhoneNum=8834834383, Password="1234", UserRole=1, CreatedDate=DateTime.Now, UpdatedDate=null, Active=true }
-            };
-            context.TblUsers.AddRange(user);
-            context.SaveChanges();
-        }
-
-        public void Dispose()
-        {
-            context.Database.EnsureDeleted();
-            context.Dispose();
+            userService = new UserService(_fixture.context, encrypt.Object);
         }
 
         [Fact]
         public void GetAll_User()
         {
             //Arrange
-            SeedDatabase();
 
             //Act
             var result = userService.GetUser();
@@ -64,15 +33,12 @@ namespace BALTest.Services
             //Assert
             var items = Assert.IsType<List<UserDisplay>>(result);
             Assert.Equal(3, items.Count);
-            
-            Dispose();
         }
 
         [Fact]
         public void Check_Extist_with_CheckExtistUser()
         {
             //Arrange
-            SeedDatabase();
             var user = new Registration() { FirstName = "Gokul", LastName = "B V", Email = "bvgokulgok@gmail.com", PhoneNum = 8834834383, Password = "1234", ConfirmPassword = "1234", UserRole = 1, CreatedDate = DateTime.Now, UpdatedDate = null, Active = true };
 
             //Act
@@ -80,15 +46,12 @@ namespace BALTest.Services
 
             //Assert
             Assert.True(result);
-
-            Dispose();
         }
 
         [Fact]
         public void Check_New_with_CheckExtistUser()
         {
             //Arrange
-            SeedDatabase();
             var user = new Registration() { FirstName = "Gokul", LastName = "B V", Email = "anish@gmail.com", PhoneNum = 8834834383, Password = "1234", ConfirmPassword = "1234", UserRole = 1, CreatedDate = DateTime.Now, UpdatedDate = null, Active = true };
 
             //Act
@@ -96,15 +59,12 @@ namespace BALTest.Services
 
             //Assert
             Assert.False(result);
-
-            Dispose();
         }
 
         [Fact]
         public void Check_Correct_ConfirmPassword()
         {
             //Arrange
-            SeedDatabase();
             var user = new Registration() { FirstName = "Gokul", LastName = "B V", Email = "bvgokulgok@gmail.com", PhoneNum = 8834834383, Password = "1234", ConfirmPassword = "1234", UserRole = 1, CreatedDate = DateTime.Now, UpdatedDate = null, Active = true };
 
             //Act
@@ -112,16 +72,12 @@ namespace BALTest.Services
 
             //Assert
             Assert.True(result);
-
-            Dispose();
         }
 
         [Fact]
         public void Check_Wrong_ConfirmPassword()
         {
             //Arrange
-            SeedDatabase();
-
             var user = new Registration() { FirstName = "Gokul", LastName = "B V", Email = "anish@gmail.com", PhoneNum = 8834834383, Password = "1234", ConfirmPassword = "4321", UserRole = 1, CreatedDate = DateTime.Now, UpdatedDate = null, Active = true };
 
             //Act
@@ -129,16 +85,12 @@ namespace BALTest.Services
 
             //Assert
             Assert.False(result);
-
-            Dispose();
         }
 
         [Fact]
         public void Test_Register()
         {
             //Arrange
-            SeedDatabase();
-
             var user = new Registration() { FirstName = "Gokul", LastName = "B V", Email = "gok@gmail.com", PhoneNum = 8834834383, Password = "1234", ConfirmPassword="1234", UserRole = 1, CreatedDate = DateTime.Now, UpdatedDate = null, Active = true };
             encrypt.Setup(method => method.EncodePasswordToBase64(user.Password)).Returns(user.Password);
 
@@ -147,15 +99,12 @@ namespace BALTest.Services
 
             //Assert
             Assert.True(result);
-
-            Dispose();
         }
 
         [Fact]
         public void LogIn_with_correct_mail_password()
         {
             //Arrange
-            SeedDatabase();
             var user = new TblUser() { Email = "bvgokulgok@gmail.com", Password = "1234" };
             encrypt.Setup(method => method.EncodePasswordToBase64(user.Password)).Returns(user.Password);
 
@@ -164,15 +113,12 @@ namespace BALTest.Services
 
             //Assert
             Assert.True(result);
-
-            Dispose();
         }
 
         [Fact]
         public void LogIn_with_correct_mail_wrong_password()
         {
             //Arrange
-            SeedDatabase();
             var user = new TblUser() { Email = "bvgokulgok@gmail.com", Password = "4321" };
             encrypt.Setup(method => method.EncodePasswordToBase64(user.Password)).Returns(user.Password);
 
@@ -181,15 +127,12 @@ namespace BALTest.Services
 
             //Assert
             Assert.False(result);
-
-            Dispose();
         }
 
         [Fact]
         public void LogIn_with_new_mail()
         {
             //Arrange
-            SeedDatabase();
             var user = new TblUser() { Email = "diptesh@gmail.com", Password = "4321" };
             encrypt.Setup(method => method.EncodePasswordToBase64(user.Password)).Returns(user.Password);
 
@@ -198,15 +141,12 @@ namespace BALTest.Services
             
             //Assert
             Assert.False(result);
-
-            Dispose();
         }
 
         [Fact]
         public void Forget_Password_Test()
         {
             //Arrange
-            SeedDatabase();
             var user = new Registration() { Email = "bvgokulgok@gmail.com", Password = "4321", ConfirmPassword="4321"};
             encrypt.Setup(method => method.EncodePasswordToBase64(user.Password)).Returns(user.Password);
 
@@ -215,8 +155,6 @@ namespace BALTest.Services
 
             //Assert
             Assert.True(result);
-
-            Dispose();
         }
     }
 }
